Experimental GraphQL.Net related changes.

+++ src/TypeCache.GraphQL/Data/PageInfo.cs
@@ -6,13 +6,13 @@ namespace TypeCache.GraphQL.Data;
 
 [GraphQLDescription("Information about pagination in a connection.")]
 public readonly record struct PageInfo(
-	[GraphQLDescription("The first cursor value of the current page.")] int StartCursor
-	, [GraphQLDescription("The last cursor value of the current page.")] int EndCursor
+	[GraphQLDescription("The first cursor value of the current page.")] long StartCursor
+	, [GraphQLDescription("The last cursor value of the current page.")] long EndCursor
 	, [GraphQLDescription("Whether a page exists before the current page.")] bool HasPreviousPage
 	, [GraphQLDescription("Whether a page exists after the current page.")] bool HasNextPage)
 {
-	public PageInfo(uint offset, uint fetch, int totalCount)
-		: this((int)offset + 1, (int)(offset + fetch), offset > 0, (offset + fetch) < totalCount)
+	public PageInfo(long offset, long fetch, long totalCount)
+		: this(offset + 1L, offset + fetch, offset > 0, (offset + fetch) < totalCount)
 	{
 	}
 }
